feat: convert 10 public marketing pages to server components with SEO metadata (audit H2)
pricing, deals, contact, products, services, wireless, connectivity,
cloud, 5g-deals, enterprise: each page.tsx is now a server component
exporting page-specific metadata (title, description, openGraph,
canonical), with the original client component moved byte-for-byte to a
colocated <Name>PageClient.tsx. Visual output and behavior unchanged.
The contact page re-exports offerEnquiryPrefix so its existing test
import path keeps working.

These pages previously could not export metadata at all ('use client'),
leaving the main marketing surface without per-page SEO tags.
